Fix PropEq to keep variables
Join constrain get and set scheme prop to single updateSchemeProp:
This greatly simplifies the code. In addition, it may be necessary to handle
type variables within the arguments

Support multiple leafsFromMeta in TypeeBuild to handle TopTypes such as id that
are not passed instead of splitting it at the typecheck level.