Make TypeCheck fromMeta specify bound type
For various options, the bound is not always an upper bound.
For arguments or constants, it is an exact bound.
It may also be a lower bound, although no examples come to mind.

For now, I am leaving all usages as upper bound to simplify this commit. Later
ones will convert segments one by one to use other bound types