Preserve argument meta when parsing the type
Earlier, the meta for both vars and arguments was replaced with a constant
TopType. This is needed within the TypeChecker because when an argument has a
value that is a TypeVar, it is put as a variable inside the Type instead of
using the Scheme SVar. Instead, the replacement is made within the Encode so it
has a smaller blast radius. This is necessary to support class data objects
where the information is not available elsewhere.

In Desugar, it should replace type var mappings such as `Just<$T=$T>` with
`Just<$T>` for the object. The mapping only makes sense within the context of
the classmap.

As part of this, the ConstList was removed from the standard library. Due to the
changes, it has now revealed an issue where replacing the class with it's object
constituents forms an infinite loop with recursive class definitions like this
one. I will remove it now as it is not being used and will add it back in a
later commit.