Cons List
This adds a Cons list to the standard library.

Update the class names within the classmap within the desugar typeNameToClass
path as well. It can use classes internally.

hasPartial add additional check for (typePartial hasPartial classType).

When building a tupleApply, it now supports multiple destination types for an
expression due to different possible classes.

Add support for TypeGraph reachesPartial with a class partial