Replace args in evaluation up front
During evaluation, args are now replaced when first calling an arrow or
performing a match rather than waiting until the moment of use. This means that
the args don't need to be propagated as an environment which simplifies the code
a bit. Most importantly, it can run into problems when multiple levels of
functions are called at once and it is unclear which level is being used.