Make list generic
Fix the hasType to allow "forgetting" a type variable, but not any other
difference in type vars.

Add a new case to encoding meta for handling m.$T = $ObjVar. These requires
setting an equivalency with a new constraint type for VarEq. This is pretty
similar in implementation to the existing PropEq.