Expr TupleApply adds one arg at a time
First, this better clarifies the order of application by ensuring that the first
items are applied before the later ones. With a map, it could be altered.

It also means that there is no difference between a(b, c) and a(b)(c). The first
is just syntactic sugar. By aligning the two, it reduces the number of possible
areas of ambiguity where conversions between the equivalent forms are necessary.

It also serves as a prerequisite to supporting tuple apply with an unknown
argument name.