Handle arrow result with expr as filter properly
With an arrow result an an expr, the result is not all possible evaluations of
the expression. So, it is neither an upper nor lower bound that is entered. What
the user actually types is that a resTreeCut should be a supertype of what they
enter. This separates what they enter (which filters the value) from the upper
bound that is expected to be returned.