Working argument inference
This finishes adding support for argument inference. It uses the type graph to
lookup all possible args when adding one as various options. Then, it ensures
only a single argument possibility is left. The current decoding system may be
unnecessarily strict, so I may loosen it in the future if necessary.

Update the working examples to use argument inference. Many of the remaining
examples do not work when converted to argument inference.