TypeGraph reaches with part of src partial
Especially early in typechecking with the partial is not narrowed down, only
some of it will go through a reaches tryArrow. It used to require that the whole
partial would fit, but that was not correct.

Remove the intersect type requirement that the type variables match. Instead,
the correct semantics is that a type variable can simply be forgotten.