Add template support for type properties
This adds most of the work to support type properties in the language. However,
it is only a template because there is no parser or any way to actually encode a
type property. The code is therefore largely untested. An actual usage of type
properties is left for a later usage.