Simplify typecheck encode with typevar substitution
The previous system was overly complicated and resulted in the loss of type vars
that should have been propagated through to decoding. Without these vars,
relationships were lost necessary to reduce the output type based on the input type.

The substitute vars can also be used later for applying the var relationships
during the type constrainment as well.