Create mainb for CatlnResult build
Instead of building only an executable, main can be used to describe building
larger constructs. This PR creates a simple web example of a build construct
along with some mechanism to test these builds. Currently, main is used for
executable and mainb for CatlnResult, but I will rename them later so main is
used for build and mainx for executable.