Start codegen
This creates basic code generation inside of Emit. The code generation does not
produce anything that is close to correct and has a number of massive issues.
The goal is to build a starting point from which infrastructure could be built
around it. For example, a WebDocs to display the completed LLVM bytestring with
syntax highlighting is the next task. After that, I can view the results to
start improving the correctness of the emitted code. In addition, I will move
the codegen execution into an Eval.Runtime arrow to unify it as part of the
CatlnResult build system.