Webdocs runs on request
Before, it would run all of the processing up front and then serve the results
on request. Now, it runs the processing on request. The first benefit of this is
that if it fails during a stage, it can still return earlier stages. In
addition, it also makes it faster to update the code and view the results
because it doesn't require restarting the server.