Set props with type varibles
When setting a property equal to a type variable, it should replace with the
type variable in the super object. This enables the representation to get passed
on to the parent.

To support this, there is a union of type and type variable in the
unionReachesTree. In this case, I currently have it set to keep the types if
there are any and the type variable otherwise. However, the correct definition
may require that Type is modified to support a sum including both partials and TypeVars.