Try to fix intersectVars
Rewrite arrowDestType to use a simpler form of the union/intersection of the
arrow and root expression. However, it requires using the argument when it is
passed an argument. I am not sure if that is quite correct, but it seems to work now.

Pass the showObject into the arrowDestType to include the current scheme and
handle it in a more systematic manner.

Move the type variables into the disabled test directory while creating
non-variable versions that do seem to pass.