Separate TreeBuild build ResExEnv
The ResExEnv does not have to be part of the makeBaseEnv stage. Instead, it is
logically clearer to have it as a separate stage as it relies only on the
baseEnv.

In addition, it may be necessary to later use the buildArrow component of the
buildResExEnv in isolation when building arrows which are not clear from the
offset of the program.