Eval Env as record
By converting to a record, it is possible to move more data into the eval env
including stack traces, coverage, debug outputs, etc.

It also separated the getArrowTree method to rely on the Env only. This will be
used later to build the arrows only when needed rather than all upfront. Besides
the gains in not building unnecessary arrows, it is more important because it is
not possible to know up front what needs to be built when dealing with macros.
This allows the macros to be built when necessary.