Measure Arrow Coverage during Eval
This measures how many times an arrow is called during evaluation. As each arrow
is equivalent to a basic block, this is essentially the branch coverage. For
testing, it can be used to measure the test coverage. For now, it serves as a
good start to using the eval environment.