Eval builds arrows as needed
As it is impossible to know what arrows need to be built ahead of time when
macros are involved, this refactor makes the evalEnv build as needed.

Included are a few small refactors. It splits the TreeBuild's envLookup into a
findResArrows method to handle part of the complexity. In the Eval, it removes
the evalBuildMain that used to be used for debugging as a better debugging
method should be possible soon with WebDocs. It also creates a buildRoot method
in TreeBuild which should be used in the future instead of buildPrgm, but
buildPrgm won't be removed until it is removed from Emit.