Treebug data collection
For debugging with webdocs, I created the treebug system. It collects a record
of the program execution to display as an evaluation tree. It collects when
starting and stopping the execution of an arrow in evaluation and collects the
data inside the eval env. To return the data, an evalResult object was created
that returns from eval. It is used in the webdocs to add a new path.