Run emit codegen with arrow
This commit enables the emit to be run using an arrow from evaluation build.

As part of this, it uses a new LLVMVal to hold the result (should be temporary).
This is returned by making the ConstantArrow hold a real Val, instead of a
Constant. In order to support this, the ResArrowTree and related code was moved
from Syntax to Eval.Common. This Emit.Common was renamed to not clash with
Eval.Common. To avoid a circular dependency, the Eval environment helpers were
all moved to a new module Eval.Env.