Replace Val in Eval with full tree
Before this, data was passed in eval using the Val from one part of the
ResArrowTree to another. This is very unintuitive resulting in bugs and
difficulty debugging. Instead, it was replaced with a straightforward use of
trees where each tree also contains the input tree. This makes the evaluation
straightforward, although it contains duplicate trees. The duplicate trees could
be resolved with a CSE pass or some kind of evaluation caching.

As part of this, I also switched from storing the arg lookups from
replaceTreeArgs into Env{evArgs}. It fits better with the rest of the system and
makes it a bit simpler to understand.

As part of this, I also fixed a bug with the conditions. When evaluating the
ifGuard, it should have the argEnvironment of the object under consideration,
even though it is outside the ResEArrow. I added additional arguments to the
ResArrowCond to include the input to the condition and build the arrArgs for
running the guard. As part of this, the Match is simpler as it no longer
requires any arg replacement (it never should have, but that was an incorrect
fix that happened to work).