Separate codegen component(s) from final result
The LLVM () monad contains the state of llvm generated so far. So, I will use
that as part of a Value so that it can be added on with additional arrows. The
final stage of generating with the LLVM will be done as part of producing the
CatlnResult as the final task.