Meta not comparable and ObjMap as list
The meta contains additional miscellaneous data such as positions that make it
unsuitable for comparing so I removed the compare type from it. As part of this,
it means that it can't be hashed so it is not possible to create an object map.
I converted the ObjMap into a list instead of a hashmap because I don't believe
I actually use it for key lookups anyway. For efficiency, it can have another
layer of indirection to lookup the object name and then the objects with that name.