Build annot definitions
Annot definitions behave very similarly to object definitions and can be built
using the same infrastructure. The necessity of annot definitions is to
determine what types to implicit the annot args into.

This PR only includes parsing and building for the annotations. I will actually
make use of the built annotation object in a later PR.

This PR also fixes a bug to accept files either with or without a trailing newline.