Run LLVM Build inside a macro
Instead of the llvm build running immediately on the function call, it sets up a
queue of functions to build with the MacroArrow and LLVMQueue. Right now, the
queue just builds the only element in it as the main codegen. Later, I will make
the queue to track subsequent arrows to build while it is building the current
arrow.

It runs by returning a MacroArrow specifying the function to run. The correct
thing would be to return a Val and let the type engine find the correct macro.