Treat Annotations as an Expression
This helps simplify the handling of the annotations. They can use objects as a
destination for building and storing definitions of annotations. It also removes
the special code needed inside the treebuild. Lastly, it can take advantage of
some of the more advanced features of expressions such as argument inference for
now and variable arguments later.