Expand reference to single partial
When specifying a data object, expand it to represent the full tuple with args.
This fixes some cases where checking for arrows with these result objects does
not necessarily behave like checking for without args.

To support this, it adds a MetaType to the MapMeta to differentiate when this
should be used vs not. The MetaType is likely to be useful for other cases later
as well.