PartialType as Record and PartialTypeArgMode
Switching the PartialType to a record for potentially more readable code as it
keeps becoming a larger and larger tuple. I added the ArgMode to describe
whether the args in it should be considered all the args or whether it should
match against potentially more args. This should help resolve some usages of
data objects as references.

Due to errors, I am only using the PtArgExact at the moment. Later, I can figure
out where the PtArgAny should be used.