Fix buildExprImp with hasPartialWithObj
The buildExprImp used to check for type equality to know whether to implicit.
However, all it really needs is to know whether the expr type is a subset of
what the desired destType is.

To fix this, I updated the hasTypeWithVarEnv to also include an ArgEnv as well
since that was needed. I also created a helper hasTypeWithObj which can be used
in a few places. I then used this helper in a few additional places.