Unify WebDocs Val handling
This makes both the build page accept a Val (from the Haskell end) and show the
result with a Val. It also combines the Common.js Val that builds CatlnResult
specific Vals with the more standard Val displays used in the Debug. Right now,
all pages use the unified Common Val so there is no duplication.