Fix id with variables
This PR fixes several bugs when working with type variables.

In the TreeBuild, it must make sure that it does not take a argument that is a
type variable and remove it from the varEnv. So, this checks for the variable
and instead does a lookup at that point.

The second bug was in the hasPartialType. When finding a new varEnv and argEnv
to recurse with, it should use the superType as opposed to the subType. The
superType is somewhat more general so it should be better. However, it may
instead require the two types to be unioned instead.