Verify scheme with ub narrows
This adds another aspect to the scheme verification that ensures that the new
scheme is smaller than the old one. It identified a few problems.

First, when checking partials, I am not counting the addition of new vars as an
issue. Vars, when both present, must be equal. And, the arguments resulting from
the vars must be right. However, the vars do not necessarily have to be there.

Following this, the recursive venv in hasType can now come from either the
superVars or the subVars so they must be combined.

Lastly, the intersectVars used to propagate the vars all the way up. Instead,
the vars in a sum can only be from the sum except for the varValues which can be
from the super venv. This corrects it to match the semantics used in the rest of
the program.

This also fixes the metadata printing in ListProgram for expressions.