Webdocs show bad typecheck with constrain
Before this, the constrain did not have a bad type check as it was only the
error. This makes it collect all the data anyway so it can be debugged.

Instead, it adds a failOnNotes so later steps can be blocked by error notes.

This also fixes the display in WebDocs of error notes and adds a notes display helper.