[WIP] Emit start on using actual results
This is a start for emit using hardcoded results instead of hardcoded zeros.
Eventually, the zeros become incompatible with more implementation. After this,
it works for some of the examples but not all of them.

This makes a number of changes:

Uncomments out all of the code to return the real results instead of hardcoded ones.

Int32 is used everyone to avoid incompatibilities in the integer types. Longs
can be inserted later along with the Catln type.

The instr command now requires the result type instead of hardcoded double.

Requires a destType for codegenDecl to specify the return value.