Fix bugs in Emit
This fixes several miscellaneous bugs that were found thanks to the
llvm-hs-pretty codegen.

First, alloca was not correct. The result of an allocation was a register
containing a pointer to the memory address. This should be used as a pointer.
Afterwards, the data must always be loaded from the address.

I added the DeclInput to the arrow hash to deal with the different calling conventions.

When calling into a function, I should pass args as they are, not by
destructuring with buildArrArgs. Instead, they should be destructured inside the
function instead.

The getelementptr must have a pointer address, and the indices contain only indices.

After main, it should load the value of initIO. It also should always return 0
from main.