Functor definition and instances with vars
This adds a basic definition for functor and two instances: option and conslist.
It adds the capabilities for instances to also include typeVars to relate to
their classes typeVars. It also adds the (untested) feature for different type
between the class and instance var names. Finally, there is a fmap function
which is for now commented out.