Improve parsing with statement newlines and remove try
This fixes the parsing so that it will parse the statements for fmap on a list.

The first major change is how newlines are handled. It provides a
pNothingNewline that is a peer to statement. This is instead of having statement
absorb empty newlines after it. I believe it should make the parser more resilient.

It also removes various try statements which are not necessary. Try involves
backtracking so it only needs to be placed where backtracking should happen.

It removes the old pValue since that can be unified with the pCall.