Parse files to graph
This changes the file parsing to produce a graph instead of a visited tree. This
works better to handle both files visited by multiple parents and dependency
cycles. Later, it can be used to typecheck only a single file at a time and use
fully typechecked dependencies as part of it.

It also creates a starting utils file to contains simple Haskell utilities.

Lastly, it updates the webdocs file display as the input format changed. Now, it
is able to use the file path as the key to locate a file which will be useful
for linking to source. It also shows the Table of contents based on the file
paths, not based on dependency loading.