PlayButton in docs declarations
This adds a green arrow button next to declarations to execute the declaration.
The execution can, at this time, be either build, debug, or constrain. Note that
it executes main, rather than the clicked declaration.

In the future, the contextualization would be more necessary. You could build
any declaration, execute certain declarations, debug it, or show the type
constrain process for it.

This then removes the global links from the header.