Desugar to prgm graph
Instead of desugaring while merging, it keeps files as a separate prgm graph.
This will be useful later for typechecking each file separately and caching the results.

This moves the logic of inserting the automatic core dependency into the parser.
It needs to be added to the import list to avoid messing up the import tree.

The prgm merging code was moved into the more general Syntax.Prgm. It can be
used for any Prgm, not just DesPrgm. It will be needed later for typechecking.

It also updates the passes in desugaring to use the full prgm containing this
and the recursive imports. This is because various dependencies will only be in
imports. This might be a bit inefficient, so could change later.