Typecheck using GraphData of files
This change enables the typechecking to be a gradual process. Most importantly,
it sets the foundation for later file-based caching of typecheck results. It
also means that different values can be within the same webdocs, but not part of
the final result by not including them in the final dependency graph.

It also opens up a bug with using functions defined in a parent function. I will
just leave it for now. The issue can be found in the list.ct example.