Show docs one page at a time
Before, all of the docs pages reachable throughout the program are built and
loaded at the same time. This makes it so that a separate call is used to load
the table of contents, and then an inner page shows the current doc. Only the
current doc is built and loaded.

This splits the docs into a main Docs and an inner DocPage. The main only
handles the table of contents and routing. The inner covers the display for the
main section of the page.

Likewise, it splits the /pages into a toc and a page request in webdocs.