Typecheck used in docs
This adds typechecking data into the webdocs docs pages. The typechecking is
used to correct usages of types into classes where necessary.

This also improves the pages syntax highlighting. It adds a keyword and typevar
to bold those elements and uses typenames and class names more liberally. Too
liberally as they are used in some places they are not supposed to and some
usages are wrong.

It moves the data, class, and instance declarations into headings instead of
divs. These are often lines that separate different areas of docs or code so it
makes sense to treat them as headings.