Webdocs caching
This creates two modes for webdocs: live and cached. The live is mostly for
local development and will eventually be sped up as more caching is added to the
compiler. The cached version is for serving on a website to reduce load. Before
the real caching is implemented, this shortcut will help make the stack webdocs
reasonable. Even after, it will still be useful to speed it up.