Fix docs page numbering
The numbering is currently based on alphebetical ordering, except main is the
first in a directory. The ordering was not being shown in the next/previous page
buttons until now.

It moves the pagetree creation into two steps. The first determines the basic
files and the second puts them into a more full representation that is also
sorted. It also creates an artifact with the sorted list at this time.