Rename "instance _ of _" to "every _ isa _"
This improves the expansion of classes to be more readable. While the statement
is not particularly clear in the action that it is taking (expanding the isa
class), it is much better in terms of describing the effect that it has. It also
works much better with more advanced usages with adding an entire class or where
it is extending only part of a class.