Add method syntax sugar to parser and webdocs
This adds a syntax sugar to define methods with the this keyword. It will make
the usage of methods more common which helps with chains of calls.

This adds both a parser for it's input and displaying code in webdocs. It
desugars immediately on parsing rather than the desugar pass. This means that
formatting should automatically convert to method format.

Finally, it is used throughout the tests and core.