Change comment from "// " to "# "
There are a few benefits from this change. Historically, it follows more of the
unix tradition rather than a C tradition and this language really doesn't follow
C much. It also requires only a single character rather than two. Combined with
the following space, it is an increase of 2 in the indentation rather than 3.

It also fixes some potential ambiguities in parsing. There may be some
significance in how it relates with annotations as well as they both are based
on the "#" symbol.

Lastly, it may be necessary to free up operators for later. Right now, "/" is
used for division. But, I might later use "//" for division so that "/" can be
used for module names. It is especially necessary for module names as it is a
single character with that same meaning and doesn't require a shit to type.