Combine standard tests and build tests
This creates a new function containsMainx to test whether mainx exists in a
prgm. Now, the default test behavior is to evaluate mainx only if it exists,
otherwise do the normal build. This means that the build and test directories
can be unified which simplifies the test code a bit.