Working catln doc command
This has the "catln doc f" now include the static files for webdocs as well. So,
it is the only part required for the working docs.

In addition, it also fixes the docs with the change from object into a record
and adds the "/api" prefix to all of the webdocs api commands.