Comments in declarations and definitions
This adds support for a comment statement inside a declaration or definition.
These comments can later serve as documentation comments to be included in the docs.

It also combines the declSingle and declTree into a single statement. This is
because even single line declarations can have comments inside them (and annotations).