Support annotation inheritance
This transforms statements into a tree so that annotations can be inherited. It
will be common to have multiple entities that share an annotation such as the
multiple test definitions updated. In this case, there can be one annotation
line and then all lines inside of it receive the annotation.