Improve object bounding
This makes two improvements to the object bounding.

First, it makes the encode for arguments bound by all objects instead of just
type objects. This is necessary for sub-objects that might be created with context.

The second improvement is that when computing the union of all objects in the
typegraph, it also considers objects which are only defined as pieces of another
object. This should be included as it still can be created as a value. With
context, this situation will also become very common.