Fix webdocs type infer links in constraints (#32)
* Show annotations in webdocs list program

This makes the webdocs also show the annotations for an arrow.

It also hides the lines showing return values when there is no information about
the return value to display.

* Fix webdocs type infer links in constraints

The links relied on finding the prgmName from the params. However, the params
are changed as it is inside not the top-level route, but a second-level route.
So, this changes to passing the prgmName to the pnts through a context object
instead.