Fixes for webdocs
This fixes the display of error messages in webdocs. It also cleans up the type
inference page styling.

Lastly, it reverses the order of the traces in an epoch. Before this, the epoch
ordering was correct(ed) but the order of constraints applied in the epoch was
reversed. This ensures both are in chronological order by the constraint process.