Refacter typing operations
This refactors many of the operations in syntax.types to make them simpler and
more reliable. It uses a standard splitPartials with consistent handling of
expandingClasses. However, it makes fewer assumptions which makes this slower
than the previous version.

It is necessary, however, because it resolves several problems. Specifically, it
fixes support for classes which contain other classes inside them. As part of a
later effort, it should be possible to improve the performance.