Rename buildUnionObj to addUnionObjToEnv (#49)
* Remove unused FEnv.boundTypeObjs

It was originally created to be similar to the boundAllObjs except contain only
types, not functions. It would be used for arguments. However, as arguments can
themselves be functions it turned out to not be needed at all.

* Rename buildUnionObj to addUnionObjToEnv

This does some minor simplification by removing old code and cleaning up
unnecessary pieces from the typegraph union obj.