Filter unionObj by precedence (#50)
This adds a scheme when constructing the Any object to only respect the highest
precedence objects. For example, if a data type is defined, that should remain
correct regardless of any function objects sharing the same name. Otherwise, a
definition or match that doesn't put as much detail into the argument types
would accidentally increase it to TopType.