Check pnt updates all in set Descriptor
This takes better advantage of the set descriptor method in env.common by moving
all checks for whether a scheme has changed there. This lets executeConstraint
and several other methods in Constrain become simpler.
Constrain

One other change is how the runLimit reached error is handled. Previously, it
ran another epoch to see what was still changing. This modifies it so that now
it simply prints out the latest trace of an epoch.