Fix overlapping partials
Before, it checked only against partials in an ordered fashion. This would only
work if the partials were sorted and subPartials must come before the
superPartials. However, this is not guaranteed so it would fail to compact when
they were in the opposite order.

Following this change, it will compact regardless of the order.