Convert ClassMap to ClassGraph
The ClassGraph allows many of the graph operations to be used within the type
operations. Although they are not entirely accurate lacking access to classVars,
they can be much faster and serve as a "pre-test" to hopefully speed up the type
operations. Especially if the graph checks can be memoized.