Add PRelativeName
This creates a new form of partial name PRelativeName. The reason it is
necessary is that a relative name can be either (or both) types and classes, so
it needs a separate option. This also simplifies some implementation as both
PTypeName and PClassName can refer only to absolute names while PRelativeName is
only relative.