Move annotations out of arrow
The arrow is both optional and represents the output for a pair. Instead, the
annotations are really more part of the input. The proof of this is that
annotations can be attached even to a type declaration which is lacking an
arrow. So, this moves the annotations to the object map as a peer of the object
and arrow