Move removeSubDeclarations to use argMetaMap
Before, it used the applied args which is not right and would require matching
again in the nested statements. Instead, it is better to just transmit the
already matched args. This also allows the argMetaMap to be fixed.

As part of this, it creates the exprArgs to eventually move to objExpr and
deprecate formArgMetaMap. However, it is currently broken.