Create desugar preprocessors for match
This creates a new system for declaration preprocessors located in expressions.
It can then move the standard multi-line expression behavior to an expresssion
pre-processor, but also leaves open possibilities for new ones. Case and
ifThenElse are the first. But, in the future ones for tables, diagrams, etc.
could also be created systematically.