Convert desugar to ExprObject
This converts the desugaring process to use ExprObject rather than old
Object (as part of exprObject).

As part of it, it also moves the parsing of types to using expressions. This now
requires changing the pExpr to use a enum mode rather than a boolean.

For now, match and case will no longer work. In order to work cleanly, it
requires adding both the AliasExpr and support for expressions as arg names.