Start using (name: Type) instead of (Type name)
This change is used to make the syntax far more systematic and easier to parse.
It can use expressions for the type that are then parsed into types later.

In that sense, it uses `operator:(l: $T, r) -> $T` as it returns the first
argument l.

Following this, a comparable version with `operator->` can be added which can be
used to support the arrow hints used in higher order functions.