Types using square braces
This converts the types to List[Int] from List<Int>. The main reason for this is
that there is difficulty parsing the < from the type variable vs the comparison
operator. Instead, the square bracket only has lists as the other usage.

As a result of this, it won't be possible to have the access operator map[key].
Instead, it may be worth doing something like map.(key). It isn't quite as nice,
but should still be fine.