Separate TypeGraph from TypeChecking
The TypeGraph system uses a ObjectMap to apply across a Type and produce other
possible types. Although it was mainly used for TypeChecking, it may be used
elsewhere including for refinement types. For that reason, I made it more
generalized and split it off. There does seem to be some performance impact due
to repeatedly building typegraphs. This can be improved later.