Parser improvements with pArrowFull
This makes a handful of parser improvements. The main one is the addition of a
pArrowFull to combine much of the logic of "input guard -> decl = def" found in
both the declaration and argument sections to help unify them.

Besides that, it also removes much of the remaining pType and pSingleType code
in favor of parsing expressions and then desugaring to type.

Finally, but also quite important, it fixes the newline handling for statement
trees. This removes the try and will greatly improve the quality of feeddback
during syntax errors.