Start making constrain use actual vs required
Right now, it uses an upper bound and an almost entirely unused lower bound.
Because the lower bound doesn't really make sense.

Instead, both are upper bounds. But, one is an upper bound that covers what the
type actually is and the other is what it should be. Because, what "should be"
doesn't change what it actually is.