Create RawObjArr
This creates a new syntax RawObjArr to handle various expansions on ObjArr.
Right now, it is just default values. In the future, it should also support
multiple inputs to same outputs or currying.

It also creates a shared ObjArrClass for them both.

This is only an initial piece to unblock the parsing and formatting of the .ctx
examples and align work to fix webdocs. Future work to complete this effort
include parsing defaults and actually having the ability to desugar RawObjArr.
This is especially true when it desugars one RawObjArr to multiple ObjArr.