Remove $ from TypeVar names
This removes the prepended $ that comes before a usage of type vars. This is for
several reasons. First, it helps deal with the variation between the internal $T
and the external $_T. Second, it will serve as progress towards eventually
unifying vars and args such that $T can eventually just be T.