Combines oaM into arrM
This puts the oaM into an optional part alongside the arrow meta. This
recognizes that there should be an arrow meta if there exists an arrow
expression, but should also be one if it is just a declaration. For type
objects, it changes it so that there will not be one rather than there being one
but it is topType.