Removes GuardExpr
This removes the GuardExpr construct in favor of having (x | y). This enables
more of it to follow existing code paths for more standardized parsing, type
checking, compiling, etc.

It instead replaces it with a new expr type called EWhere.

Once it is working, the same will be done for the ObjArr as RawObjArr. During
type checking, the reachesPartial can be modified with a special case to handle
operator|. It would modify the source with a condition. Then, the TreeBuild and
Eval sections would rely on the type of the objExpr.