Move TypeVarLoc to TypeVar from TypeVarAux
This makes it so that the location of the variable is part of the Type rather
than the name of the variable. This is because the usage of it is not universal.
For example, maps such as the VarArgEnv don't have an internal/external
separation, they are just the values. Only the actual usages within the Type
context have it as a property.