Adds missing apply BoundedByObjs, improves tb errors
This adds some BoundedByObjs into the input tupleApply. It also improves the
error messaging for failing to buildGuardTrees. As part of this, it adds a
recursive error messaging system that can be used to actually understand errors.
It also means less problems are swallowed by the TB handling. Using this,
several existing bugs in stack code have been fixed.