Makes arguments as implicit result
This is a large refactor. Rather than having arguments have a type, they behave
similarly to existing language structures such that they are their own type with
an implicit with the replacement of the arg. This should also behave much better
for higher order functions. As part of this, numerous other changes and
refactors occurred:

The Expr Arg was removed. It can now be unified with Val as they are handled
largely the same. What logic there was is now part of the exprVarArgs and
family.

Various places in desugar, expr buiding, etc. had to be modified to make the
type of the argument as the arg arrM rather than the objM.

For argument matching, the arguments are now added as part of the ReachesEnv.

The syntax was changed. Now, ":" is used as an alias for -> while :: is instead
used for type operations. I am not sure how this will change in the future, but
the basic issue is -> is annoying to write, : is less clear, and :: is fine as
it is not used too often.

This verifies that the encode clearMetaArgs is no longer required and removes
it. Doing so improves the quality of several typecheck results that would
otherwise result in TopType. It also offers no benefit.

A new integer identifier is attached to constraints. This purpose is to allow a
constraint to be matched up to the location in the encoder.

This finishes most of the rest of the changes in moving from Object to objExpr.
The last pieces in encoding and TypeUtils topObj can now use expressions. In a
later PR, the final changes and removal can take place.

The constraints were combined into a Constraint with generic. This is used in a
few places now, but may be more later. One potential plan is to simplify the
constraint, loading, and saving into three seperate and much simpler operations.

The Eval and TreeBuild system was re-written from ResArrow to TCExpr and TCCall.
The source is that as of this change, implicits can only be in clearly defined
parts of an expression. So, it can be remodeled where TreeBuild takes the
expression with implicits and makes them explicit. This also greatly simplifies
the code and makes it much clearer.

There were also a number of miscellaneous bug fixes that occurred. Some were
newly introduced, but a number of existing bugs were also scrounged up.